test/app/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-hybridrouter
drf-spectacular==0.27.2